Turn `NamedValue` into an interface replacing `NamedObject`.
Almost everything is pushed down into the two sole inheritors, `Procedure` and `SyntacticForm`. The latter is no longer a `FusionValue`, which was always misleading.
